# docs/swagger.yaml
basePath: /api
definitions:
...
  models.AddTeamMemberInput:
    properties:
      team_id:
        type: string
      user_id:
        type: string
    type: object
  models.AddTeamMemberToMyInput:
    properties:
      user_id:
        type: string
    type: object
  models.ChangeTeamNameInput:
    properties:
      name:
        type: string
    type: object
...
  models.CreateTeamInput:
    properties:
      leader_id:
        type: string
      name:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      created_at:
        type: string
      error:
        type: string
    type: object
...
  models.Role:
    enum:
    - super_admin
    - team_leader
    - team_member
    type: string
    x-enum-varnames:
    - SuperAdminRole
    - TeamLeaderRole
    - TeamMemberRole
  models.Team:
    properties:
      createdAt:
        type: string
      id:
        type: string
      leader:
        $ref: '#/definitions/models.User'
      leaderID:
        type: string
      members:
        items:
          $ref: '#/definitions/models.TeamMember'
        type: array
      name:
        type: string
    type: object
  models.TeamMember:
    properties:
      createdAt:
        type: string
      id:
        type: string
      team:
        $ref: '#/definitions/models.Team'
      teamID:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/models.Role'
    type: object
...
  models.UserShortResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
    type: object
host: localhost:4000
info:
  contact: {}
  description: RESTful API for signing and verifying files with digital signatures
  title: Tawtheeq API
  version: "1.0"
paths:
...
  /my/team:
    get:
      consumes:
      - application/json
      description: Get the team of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Team'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get my team
      tags:
      - teams
  /my/team/members:
    get:
      consumes:
      - application/json
      description: Get all users in the authenticated user's team
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserShortResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all users in my team
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Add a user to the authenticated user's team
      parameters:
      - description: Add Team Member Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AddTeamMemberToMyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TeamMember'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Add user to my team
      tags:
      - teams
  /my/team/members/{user_id}:
    delete:
      consumes:
      - application/json
      description: Remove a user from the authenticated user's team
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TeamMember'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Remove user from my team
      tags:
      - teams
...
  /teams:
    get:
      consumes:
      - application/json
      description: Get all teams with pagination
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Team'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all teams
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Create a new team
      parameters:
      - description: Create Team Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateTeamInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Team'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Create team
      tags:
      - teams
  /teams/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Team'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Remove team
      tags:
      - teams
  /teams/{id}/leader:
    put:
      consumes:
      - application/json
      description: Update the leader of a team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: New leader ID
        in: formData
        name: leader_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Team'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Update team leader
      tags:
      - teams
  /teams/{id}/name:
    put:
      consumes:
      - application/json
      description: Update the name of a team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: New team name
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ChangeTeamNameInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Team'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Update team name
      tags:
      - teams
  /teams/{team_id}/users:
    get:
      consumes:
      - application/json
      description: Get all users in a team with pagination
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserShortResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all users in a team
      tags:
      - teams
  /teams/members:
    post:
      consumes:
      - application/json
      description: Add a user to a team
      parameters:
      - description: Add Team Member Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AddTeamMemberInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TeamMember'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Add user to team
      tags:
      - teams
  /teams/members/{team_id}/{user_id}:
    delete:
      consumes:
      - application/json
      description: Remove a user from a team
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TeamMember'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Remove user from team
      tags:
      - teams
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"